+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Plugins: Show driver plugins in the list of loaded plugins
 - Plugins: Show version of Adminer next to plugins not matching this version
 - Plugins: Remove edit-calendar, tinymce, json-column, pretty-json-column, translation, email-table, dump-php, master-slave
+- Plugins: Do not fail with a fatal error when a plugin is not an object (bug TimWolla/docker-adminer#211)
 
 ## Adminer 5.5.1 (released 2026-07-21)
 - Ignore invalid X-Forwarded-Prefix (GHSA-fr74-9mf9-gf44)